requirements.txt
asgiref==3.8.1
asttokens==3.0.0
attrs==25.1.0
backcall==0.2.0
beautifulsoup4==4.13.3
bleach==6.2.0
blinker==1.9.0
Brotli==1.0.9
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comm==0.2.2
ConfigArgParse==1.7
contourpy==1.3.1
cycler==0.11.0
dash==2.18.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dataclasses==0.6
decorator==5.2.1
defusedxml==0.7.1
dj-rest-auth==7.0.1
Django==4.2.20
django-allauth==65.7.0
django-cors-headers==4.7.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.5.0
djongo==1.2.31
docopt==0.6.2
exceptiongroup==1.2.2
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.13.1
Flask==3.0.3
fonttools==4.55.3
fsspec==2024.6.1
idna==3.10
imageio==2.37.0
importlib_metadata==8.6.1
ipython==8.12.3
ipywidgets==8.1.5
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
jupyterlab_widgets==3.0.13
kiwisolver==1.4.8
MarkupSafe==2.1.5
matplotlib==3.10.0
matplotlib-inline==0.1.7
mistune==3.1.3
mkl_fft
mkl_random
mkl-service
mongoengine==0.29.1
mpmath==1.3.0
narwhals==1.28.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
numpy==2.0.1
open3d==0.19.0
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pickleshare==0.7.5
pillow==11.0.0
pip==25.0
pipreqs==0.5.0
platformdirs==4.3.6
plotly==6.0.0
prompt_toolkit==3.0.50
pure_eval==0.2.3
Pygments==2.19.1
PyJWT==2.9.0
pymongo==3.12.3
pyparsing==3.2.0
PyQt6==6.7.1
PyQt6_sip==13.9.1
python-dateutil==2.9.0.post0
pytz==2025.1
pywin32==308
pyzmq==26.4.0
referencing==0.36.2
requests==2.32.3
retrying==1.3.4
rpds-py==0.23.1
scikit-learn==1.6.1
scipy==1.15.2
setuptools==75.8.0
sip==6.10.0
six==1.16.0
soupsieve==2.6
sqlparse==0.5.3
stack-data==0.6.3
sympy==1.13.1
threadpoolctl==3.5.0
tifffile==2025.1.10
tinycss2==1.4.0
tomli==2.0.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
trimesh==4.6.1
typing_extensions==4.12.2
tzdata==2025.1
unicodedata2==15.1.0
urllib3==2.3.0
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.6
wheel==0.45.1
widgetsnbextension==4.0.13
yarg==0.1.9
zipp==3.21.0
coverage==7.4.3
